requirements/backend.txt
# Backend Python Dependencies
# Core Django framework and REST API
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Configuration management
python-decouple==3.8

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# AI Integration
google-generativeai==0.3.1

# File handling
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Development tools
django-extensions==3.2.3
